agent: monitor: Invalid up/downlink rate in Associated Link Metrics
Currently tx_phy_rate_100kb_avg and rx_phy_rate_100kb_avg is assigned to
uplink_rate and downlink_rate, it should not be the case.

tx_phy_rate_100kb_avg should be assigned to downlink rate and
rx_phy_rate_100kb_avg should be assigned to uplink rate.

For 160MHz with 2x2 configuration, while storing accumulated phy_rate values
in tx/rx_phy_rate_100kb_acc is overflowing.

Increase the integer bit values for tx/rx_phy_rate_100kb_acc from
uint16_t to uint32_t.

Closes : PPM-3478

+++ agent/src/beerocks/fronthaul_manager/monitor/monitor_db.h
@@ -106,10 +106,10 @@ public:
         int8_t rx_snr_curr             = beerocks::SNR_INVALID;
         uint16_t tx_phy_rate_100kb_avg = 0;
         uint16_t tx_phy_rate_100kb_min = 0;
-        uint16_t tx_phy_rate_100kb_acc = 0;
+        uint32_t tx_phy_rate_100kb_acc = 0;
         uint16_t rx_phy_rate_100kb_avg = 0;
         uint16_t rx_phy_rate_100kb_min = 0;
-        uint16_t rx_phy_rate_100kb_acc = 0;
+        uint32_t rx_phy_rate_100kb_acc = 0;
 
         bwl::SStaStats hal_stats = {};
